Update the step to signal that Android is ready and iOS in progress
Android is production ready on Shorebird with iOS still in dev.
This step therefore only supports Android for now and iOS is
being worked on.

Logic and params for iOS is left in tact but not executed.

If ios or both is selected for the platform, the step will fail.